requirements.txt
asn1crypto==1.0.1
certifi==2019.9.11
cffi==1.12.3
cryptography==2.7
future==0.17.1
pycparser==2.19
PySocks==1.7.1
python-telegram-bot==12.1.1
six==1.12.0
tornado==6.0.3
